再次修正請假單api

+++ TimeWaltz/WebApplication1/WebApplication1/Services/RequestStatusService.cs
@@ -21,7 +21,7 @@ namespace WebApplication1.Services
             {
                  TableType = TableTypeEnum.請假單,
                  TableId = empId,
-                 Status = (int)RequestStatusEnum.簽核中,
+                 Status = RequestStatusEnum.簽核中,
              });
             _timeWaltzContext.SaveChanges();
         }
